asdadfasdfadf

+++ SwankRatsRoboterSoftware/Robot.py
@@ -1,28 +1,27 @@
 __author__ = 'Johannes'
-#import Adafruit_BBIO.GPIO as GPIO
+# import Adafruit_BBIO.GPIO as GPIO
 import DMCC
 import time
 
+
 class Robot:
-
-
     def setLeftMotor(self, percent):
         actuel = 0
         if percent == 0:
             print('leftMotor ' + str(percent))
-            DMCC.setMotor(0, 1, percent * 100 * 0.60)
+            DMCC.setMotor(0, 1, int(percent * 100 * 0.60))
         if percent > 0:
             actuel = 10
             while actuel <= percent:
                 print('leftMotor ' + str(actuel))
-                DMCC.setMotor(0, 1, actuel * 100 * 0.60)
+                DMCC.setMotor(0, 1, int(actuel * 100 * 0.60))
                 actuel += 5
                 time.sleep(0.30)
         if percent < 0:
             actuel = -10
             while actuel >= percent:
                 print('leftMotor ' + str(actuel))
-                DMCC.setMotor(0, 1, actuel * 100 * 0.60)
+                DMCC.setMotor(0, 1, int(actuel * 100 * 0.60))
                 actuel -= 5
                 time.sleep(0.300)
 
@@ -30,19 +29,19 @@ class Robot:
         actuel = 0
         if percent == 0:
             print('rightMotor ' + str(percent))
-            DMCC.setMotor(0, 2, percent * 100 * 0.60)
+            DMCC.setMotor(0, 2, int(percent * 100 * 0.60))
         if percent > 0:
             actuel = 10
             while actuel <= percent:
                 print('rightMotor ' + str(actuel))
-                DMCC.setMotor(0, 2, actuel * 100 * 0.60)
+                DMCC.setMotor(0, 2, int(actuel * 100 * 0.60))
                 actuel += 5
                 time.sleep(0.30)
         if percent < 0:
             actuel = -10
             while actuel >= percent:
                 print('rightMotor ' + str(actuel))
-                DMCC.setMotor(0, 2, actuel * 100 * 0.60)
+                DMCC.setMotor(0, 2, int(actuel * 100 * 0.60))
                 actuel -= 5
                 time.sleep(0.30)
